app/pipeline/service-registration/caselaw/requirements.txt
#DB
mysql-connector-python
PyMySQL
SQLAlchemy

#Application-logic
beautifulsoup4
pyyaml
pandas